django_project/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-apscheduler==0.3.0
django-cors-headers==3.2.1
APScheduler==3.6.3
pygments==2.6.1
django-oauth-toolkit==1.3.2
requests==2.23.0
djangorestframework-simplejwt==4.4.0